app_spark_to_gcs/requirements.txt
pyspark==3.1.1
google-cloud-storage==2.9.0
numpy==1.24.2
